Removes the optionality on `output` in ABI docs.
I _believe_ (someone more knowledgable than I should verify this claim) that Solidity will always write out an `output` member.  If there are no return parameters, then the array will be empty, but the `output` property will be present.

If this is not correct, I will withdraw this PR, but if the assertion is correct then the docs should reflect that so people can rely on the `output` property being present when writing ABI processors.

I also fixed some inconsistency in line termination character.  It was sometimes `.` and sometimes `;`.  I picked `.`, but don't really care which as long as it is consistent.